pyproject.toml
[tool.poetry]
name = "loan-default-predictor"
version = "0.1.0"
description = "Loan default prediction — imports core abstractions from Mlops-Plumbing"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"

# ── Mlops-Plumbing pulled in as a git dependency ──────────────────────────────
# This is the KEY line — all shared abstractions (BaseModel, feature pipeline,
# evaluation utilities) are imported directly from the plumbing repo.
ml-project-template = { git = "https://github.com/rohibindal01/Mlops-Plumbing.git", branch = "main" }

# ── Project-specific dependencies ─────────────────────────────────────────────
scikit-learn = ">=1.4.0"
pandas = ">=2.2.0"
numpy = ">=1.26.0"
mlflow = ">=2.13.0"
fastapi = ">=0.111.0"
uvicorn = { version = ">=0.29.0", extras = ["standard"] }
hydra-core = ">=1.3.2"
dvc = { version = ">=3.50.0", extras = ["s3"] }
omegaconf = ">=2.3.0"
pydantic = ">=2.7.0"
joblib = ">=1.4.0"
matplotlib = ">=3.8.0"
plotly = ">=5.22.0"
requests = ">=2.31.0"
imbalanced-learn = ">=0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"
pytest-cov = ">=5.0.0"
pytest-asyncio = ">=0.23.0"
httpx = ">=0.27.0"
ruff = ">=0.4.0"
mypy = ">=1.10.0"
pre-commit = ">=3.7.0"
pandas-stubs = ">=2.2.0"
types-requests = ">=2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM", "TCH", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["notebooks/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["src/serving/__main__.py"]

[tool.coverage.report]
show_missing = true
fail_under = 80
